Implementar testes unitários

+++ desafio/desafio/Controllers/ItemController.cs
@@ -29,34 +29,49 @@ namespace desafio.Controllers
         [HttpGet("BuscarTipo/{type}")]
         public ActionResult<List<Item>> BuscarTipo(string type)
         {
-            var item = _itemService.BuscarTipo(type);
-            if (item == null)
+            Item itemVerificador = new Item();
+            itemVerificador.Type = type;
+            itemVerificador.State = type;
+
+            if (type == "CD" || type == "DVD" || type == "Livro" ||
+                    type == "Disponível" || type == "Indisponível")
+            {
+                var item = _itemService.BuscarTipo(type);
+                return item;
+            }
+            else
             {
                 return NotFound();
             }
-
-            return item;
+            
         }
 
         [HttpGet("{id}", Name = "GetItem")]
         public ActionResult<Item> Get(string id)
         {
-            var item = _itemService.Get(id);
-            if (item == null)
+            Item item = null;
+            if(id.Length > 0)
             {
-                return NotFound();
+                item = _itemService.Get(id);
+                if (item == null)
+                {
+                    return NotFound();
+                }
             }
-
-            return item;
+                return item;
         }
 
         [HttpGet("Buscar/{name}")]  // seu get vai controller/buscar/name
         public ActionResult<Item> Buscar(string name)
         {
-            var item = _itemService.Buscar(name);
-            if (item == null)
+            Item item = null;
+            if(name.Length > 0)
             {
-                return NotFound();
+                item = _itemService.Buscar(name);
+                if (item == null)
+                {
+                    return NotFound();
+                }
             }
 
             return item;
@@ -65,36 +80,60 @@ namespace desafio.Controllers
         [HttpPost]
         public ActionResult<Item> Create(Item item)
         {
-            _itemService.Create(item);
-            return CreatedAtRoute("GetItem", new { id = item.Id.ToString() }, item);
+            if(item == null)
+            {
+                return NotFound();
+            }
+            if (item.Name.Length > 1 && item.BorrowedTo == "Ninguém" && item.State == "Disponível" || item.State == "Indisponível")
+            {
+                _itemService.Create(item);
+                return CreatedAtRoute("GetItem", new { id = item.Id.ToString() }, item);
+            }
+
+            return NotFound();
         }
 
         [HttpPut("{id}")]
         public IActionResult Update(string id, Item itemIn)
         {
-            var item = _itemService.Get(id);
+            Item item;
 
-            if(item == null)
+            if(id.Length > 0)
+            {
+                item = _itemService.Get(id);
+            }
+
+            if (itemIn == null)
             {
                 return NotFound();
             }
+            if (itemIn.Name.Length > 1 && itemIn.State == "Disponível" || itemIn.State == "Indisponível")
+            {
+                _itemService.Update(id, itemIn);
+                return NoContent();
+            }
 
-            _itemService.Update(id, itemIn);
-            return NoContent();
+            return NotFound();
         }
 
         [HttpDelete("{id}")]
         public IActionResult Delete(string id)
         {
-            var item = _itemService.Get(id);
-            
-            if(item == null)
+            if(id.Length > 0)
             {
-                return NotFound();
+                var item = _itemService.Get(id);
+                if (item == null)
+                {
+                    return NotFound();
+                }
+
+                _itemService.Remove(item);
+                return NoContent();
             }
 
-            _itemService.Remove(item);
-            return NoContent();
+            return NotFound();
+            
+           
         }
 
     }
